refactor: consolidate subscription components in shared module
- Move individual and user subscription components to shared/components
- Create base SharedSubscriptionButtonComponent with shared template
- Define ButtonMode type for reusability
- Update imports to use buttonMode() method pattern
