--- src/app/domains/profile/shared/user-subscription.widget.html
@@ -1,31 +1,0 @@
-@if (isFollowing$ | async) {
-  <button *ngIf="mode === 'FAB'" mat-fab color="subscribed" (click)="unfollow()">
-    <mat-icon svgIcon="subscribe-active"></mat-icon>
-  </button>
-  <button
-    *ngIf="mode === 'BUTTON'"
-    mat-flat-button
-    color="subscribed"
-    (click)="unfollow()"
-    class="user-subscription-button__button"
-    translate
-    data-test-id="unfollowButton"
-  >
-    Abonniert
-  </button>
-} @else {
-  <button *ngIf="mode === 'FAB'" mat-fab color="primary" (click)="follow()">
-    <mat-icon svgIcon="subscribe"></mat-icon>
-  </button>
-  <button
-    *ngIf="mode === 'BUTTON'"
-    mat-flat-button
-    color="primary"
-    (click)="follow()"
-    class="user-subscription-button__button"
-    translate
-    data-test-id="followButton"
-  >
-    Abonnieren
-  </button>
-}
