+++ src/app/shared/components/feature-subscription/shared-subscription.widget.scss
