+++ src/app/shared/components/feature-subscription/shared-subscription.widget.ts
@@ -1,0 +1,9 @@
+export type ButtonMode = 'FAB' | 'BUTTON';
+
+export abstract class SharedSubscriptionButtonComponent {
+  abstract buttonMode(): ButtonMode;
+
+  abstract follow(): void;
+
+  abstract unfollow(): void;
+}
